refactor: move TagEncodingFixer from Scanners into Helpers
It's a pure utility class with static methods — not a scanner. Lives
alongside Ulid, Uuid, and the other helpers.
